devtools/server/tests/chrome/test_inspector-scroll-into-view.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=901250
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 901250</title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gInspectee = null;
let gWalker = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  gInspectee = doc;
  const inspector = await target.getFront("inspector");
  gWalker = inspector.walker;
  runNextTest();
});

addTest(async function testScrollIntoView() {
  const id = "#scroll-into-view";
  let rect = gInspectee.querySelector(id).getBoundingClientRect();
  const nodeFront = await gWalker.querySelector(gWalker.rootNode, id);
  let inViewport = rect.x >= 0 &&
                   rect.y >= 0 &&
                   rect.y <= gInspectee.defaultView.innerHeight &&
                   rect.x <= gInspectee.defaultView.innerWidth;

  ok(!inViewport, "Element is not in viewport.");

  await nodeFront.scrollIntoView();

  SimpleTest.executeSoon(() => {
    rect = gInspectee.querySelector(id).getBoundingClientRect();
    inViewport = rect.x >= 0 &&
                 rect.y >= 0 &&
                 rect.y <= gInspectee.defaultView.innerHeight &&
                 rect.x <= gInspectee.defaultView.innerWidth;

    ok(inViewport, "Element is in viewport.");

    runNextTest();
  });
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=901250">Mozilla Bug 901250</a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
